# schema/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "description": "Canonical daemon HTTP error body.\n\nWire shape: `{ \"code\": \"<machine_code>\", \"message\": \"<human text>\" }` plus an\noptional `details` object. `code` is a stable snake_case token (e.g.\n`not_found`, `bad_request`, `runtime_unavailable`, `conflict`,\n`internal_error`, `payload_unavailable`). `message` is human-readable\nEnglish; setup-v2 error classifiers and the clipboard restore-410 handler\nsubstring-match it, so the strings are LOAD-BEARING — do not silently reword.\n\n`Deserialize` is added (the original webserver struct was Serialize-only) so\nthe Rust `uc-daemon-client` and tests can decode error bodies.",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "description": "Optional structured context (per §0.3). E.g. the restore-410\n`payload_unavailable` error carries `{ entry_id, rep_id, state }`.\nOmitted from the wire when `None`.",
            "nullable": true
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "PeerSnapshotListEnvelope": {
        "description": "Canonical success envelope: `{ \"data\": T, \"ts\": <unix millis i64> }`.\n\n`ts` is `chrono::Utc::now().timestamp_millis()`, set in the webserver handler\nvia [`ApiEnvelope::now`] (the contract carries only the type + the clock\nhelper, not a hard dependency on when the handler reads the clock).\n`rename_all = \"camelCase\"` is a no-op for the single-word fields here but is\ndeclared for forward-compat.\n\nIMPORTANT (utoipa v4): every concrete `ApiEnvelope<X>` that needs a named\nOpenAPI component is declared in the `#[aliases(...)]` block below. Add a new\nalias line whenever a new payload type needs enveloping. NEVER register the\nbare `ApiEnvelope` in `components(schemas(...))` — utoipa errors on a bare\ngeneric, and an un-aliased generic inlines an anonymous schema.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PeerSnapshotDto"
            },
            "type": "array"
          },
          "ts": {
            "description": "Server time when the response was built (unix epoch milliseconds).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "ts"
        ],
        "type": "object"
      },
...
      "PresenceRefreshEnvelope": {
        "description": "Canonical success envelope: `{ \"data\": T, \"ts\": <unix millis i64> }`.\n\n`ts` is `chrono::Utc::now().timestamp_millis()`, set in the webserver handler\nvia [`ApiEnvelope::now`] (the contract carries only the type + the clock\nhelper, not a hard dependency on when the handler reads the clock).\n`rename_all = \"camelCase\"` is a no-op for the single-word fields here but is\ndeclared for forward-compat.\n\nIMPORTANT (utoipa v4): every concrete `ApiEnvelope<X>` that needs a named\nOpenAPI component is declared in the `#[aliases(...)]` block below. Add a new\nalias line whenever a new payload type needs enveloping. NEVER register the\nbare `ApiEnvelope` in `components(schemas(...))` — utoipa errors on a bare\ngeneric, and an un-aliased generic inlines an anonymous schema.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PresenceRefreshResponse"
          },
          "ts": {
            "description": "Server time when the response was built (unix epoch milliseconds).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "ts"
        ],
        "type": "object"
      },
      "PresenceRefreshResponse": {
        "description": "Result of a `POST /presence/refresh` round.\n\n主动 probe 一轮 `ensure_reachable_all` 后的统计回执。UI 不靠这里直接判定\n在线状态：probe 过程中各设备的 Online/Offline 变化会通过既有\n`peers.changed` WebSocket 链路推送，前端再走 `GET /paired-devices`\n重拉。该响应只用于调用方显示进度或排障。",
        "properties": {
          "errors": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "offline": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "online": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "total": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "total",
          "online",
          "offline",
          "errors"
        ],
        "type": "object"
      },
...
      "RestoreEntryEnvelope": {
        "description": "Canonical success envelope: `{ \"data\": T, \"ts\": <unix millis i64> }`.\n\n`ts` is `chrono::Utc::now().timestamp_millis()`, set in the webserver handler\nvia [`ApiEnvelope::now`] (the contract carries only the type + the clock\nhelper, not a hard dependency on when the handler reads the clock).\n`rename_all = \"camelCase\"` is a no-op for the single-word fields here but is\ndeclared for forward-compat.\n\nIMPORTANT (utoipa v4): every concrete `ApiEnvelope<X>` that needs a named\nOpenAPI component is declared in the `#[aliases(...)]` block below. Add a new\nalias line whenever a new payload type needs enveloping. NEVER register the\nbare `ApiEnvelope` in `components(schemas(...))` — utoipa errors on a bare\ngeneric, and an un-aliased generic inlines an anonymous schema.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RestoreEntryResponse"
          },
          "ts": {
            "description": "Server time when the response was built (unix epoch milliseconds).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "ts"
        ],
        "type": "object"
      },
      "RestoreEntryResponse": {
        "description": "Response body for `POST /clipboard/restore/:entry_id`.\n\nWrapped in `ApiEnvelope` per §0.1. The success body was previously the\nad-hoc `{ \"success\": true }` shape; consumers ignore it today.",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "SpaceMemberDto": {
        "properties": {
          "channel": {
            "description": "Phase 96 INDIC-01:连接通道 4 态 wire 字符串。同 `PeerSnapshotDto.channel`,\n取值严格限定 `\"direct\" | \"relay\" | \"offline\" | \"unknown\"`。前端\n`SpaceMember` 直接消费,`ConnectionChannelBadge` 渲染。",
            "type": "string"
          },
          "connected": {
            "type": "boolean"
          },
          "connectionAddress": {
            "description": "当前活跃连接地址。直连时为对端 IP:port,中转时为 relay 地址。",
            "nullable": true,
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          },
          "lastSeenAtMs": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "pairingState": {
            "type": "string"
          },
          "peerId": {
            "type": "string"
          }
        },
        "required": [
          "peerId",
          "deviceName",
          "pairingState",
          "connected",
          "channel"
        ],
        "type": "object"
      },
      "SpaceMemberListEnvelope": {
        "description": "Canonical success envelope: `{ \"data\": T, \"ts\": <unix millis i64> }`.\n\n`ts` is `chrono::Utc::now().timestamp_millis()`, set in the webserver handler\nvia [`ApiEnvelope::now`] (the contract carries only the type + the clock\nhelper, not a hard dependency on when the handler reads the clock).\n`rename_all = \"camelCase\"` is a no-op for the single-word fields here but is\ndeclared for forward-compat.\n\nIMPORTANT (utoipa v4): every concrete `ApiEnvelope<X>` that needs a named\nOpenAPI component is declared in the `#[aliases(...)]` block below. Add a new\nalias line whenever a new payload type needs enveloping. NEVER register the\nbare `ApiEnvelope` in `components(schemas(...))` — utoipa errors on a bare\ngeneric, and an un-aliased generic inlines an anonymous schema.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SpaceMemberDto"
            },
            "type": "array"
          },
          "ts": {
            "description": "Server time when the response was built (unix epoch milliseconds).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "ts"
        ],
        "type": "object"
      },
      "StatusEnvelope": {
        "description": "Canonical success envelope: `{ \"data\": T, \"ts\": <unix millis i64> }`.\n\n`ts` is `chrono::Utc::now().timestamp_millis()`, set in the webserver handler\nvia [`ApiEnvelope::now`] (the contract carries only the type + the clock\nhelper, not a hard dependency on when the handler reads the clock).\n`rename_all = \"camelCase\"` is a no-op for the single-word fields here but is\ndeclared for forward-compat.\n\nIMPORTANT (utoipa v4): every concrete `ApiEnvelope<X>` that needs a named\nOpenAPI component is declared in the `#[aliases(...)]` block below. Add a new\nalias line whenever a new payload type needs enveloping. NEVER register the\nbare `ApiEnvelope` in `components(schemas(...))` — utoipa errors on a bare\ngeneric, and an un-aliased generic inlines an anonymous schema.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StatusResponse"
          },
          "ts": {
            "description": "Server time when the response was built (unix epoch milliseconds).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "ts"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session_header": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "session_query": {
        "in": "query",
        "name": "auth",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Local daemon HTTP API for the UniClipboard GUI and native clients. All enveloped responses use the canonical `{ data, ts }` shape; errors use `{ code, message, details? }`. Binary and WebSocket endpoints are exempt from the envelope. L2+ operations require a session token (query `?auth=` or the `Authorization` header).",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "title": "UniClipboard Daemon API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/clipboard/restore/{entry_id}": {
      "post": {
        "description": "Re-apply a stored clipboard entry to the local system clipboard. Wrapped in\nthe canonical `{ data, ts }` envelope (ADR-008 §0.1/§0.2); errors use the\ncanonical `ApiErrorResponse`. The `payload_unavailable` (410) error carries\n`{ entry_id, rep_id, state }` in `details` (§0.3); the `code`/`message`\nstrings are LOAD-BEARING and preserved.",
        "operationId": "restoreClipboardEntry",
        "parameters": [
          {
            "description": "Clipboard entry id to restore",
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Restore as plain text only (strip rich representations)",
            "in": "query",
            "name": "plain",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreEntryEnvelope"
                }
              }
            },
            "description": "Entry restored to the system clipboard"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Entry not found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Entry payload is no longer available (orphaned/lost)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_query": []
          },
          {
            "session_header": []
          }
        ],
        "summary": "POST /clipboard/restore/{entry_id}",
        "tags": [
          "clipboard"
        ]
      }
    },
...
    "/paired-devices": {
      "get": {
        "description": "List paired space members with presence. Wrapped in the canonical\n`{ data, ts }` envelope (ADR-008 §0.2); the previous bare top-level array\nis retired.",
        "operationId": "listPairedDevices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberListEnvelope"
                }
              }
            },
            "description": "Paired space-member list"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_query": []
          },
          {
            "session_header": []
          }
        ],
        "summary": "GET /paired-devices",
        "tags": [
          "system"
        ]
      }
    },
...
    "/peers": {
      "get": {
        "description": "List discovered peer snapshots (topology view). Wrapped in the canonical\n`{ data, ts }` envelope (ADR-008 §0.2); the previous bare top-level array\nis retired.",
        "operationId": "listPeers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerSnapshotListEnvelope"
                }
              }
            },
            "description": "Peer snapshot list"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_query": []
          },
          {
            "session_header": []
          }
        ],
        "summary": "GET /peers",
        "tags": [
          "system"
        ]
      }
    },
    "/presence/refresh": {
      "post": {
        "description": "Actively probe paired peers' reachability and return the round's counters.\nWrapped in the canonical `{ data, ts }` envelope (ADR-008 §0.2); the\nprevious bare object (counters at top level) is retired.",
        "operationId": "refreshPresence",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceRefreshEnvelope"
                }
              }
            },
            "description": "Presence refresh round completed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "session_query": []
          },
          {
            "session_header": []
          }
        ],
        "summary": "POST /presence/refresh",
        "tags": [
          "system"
        ]
      }
    },
...
    "/status": {
      "get": {
        "description": "Diagnostic snapshot: version metadata, uptime, and worker health. Wrapped\nin the canonical `{ data, ts }` envelope (ADR-008 §0.2); the previous bare\nobject shape is retired.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusEnvelope"
                }
              }
            },
            "description": "Daemon status snapshot"
          }
        },
        "security": [
          {
            "session_query": []
          },
          {
            "session_header": []
          }
        ],
        "summary": "GET /status",
        "tags": [
          "system"
        ]
      }
    },
...
  },
...
}